Build without bind mount (#54)
* Refactor docker build
