+++ .gitignore
@@ -5,11 +5,6 @@
 # docker image archives
 *.tar.gz
 
-# build artifacts
-openmpf_runtime/build_artifacts/*
-!openmpf_runtime/build_artifacts/.gitkeep
-install.tar
-
 # backup
 *.bak
 
@@ -19,3 +14,4 @@ docker-compose-test.yml
 swarm-compose.yml
 openmpf-docker-images*
 .env
+test-reports
