+++ components/cpp_component_build/README.md
@@ -1,0 +1,2 @@
+The `openmpf_cpp_component_build` image works in conjunction with the `openmpf_cpp_executor` image. 
+Documentation for both can be found in [`cpp_executor/README.md`](../cpp_executor/README.md).
