+++ components/java_component_build/README.md
@@ -1,0 +1,2 @@
+The `openmpf_java_component_build` image works in conjunction with the `openmpf_java_executor` image. 
+Documentation for both can be found in [`java_executor/README.md`](../java_executor/README.md).
