+++ components/python_component_build/README.md
@@ -1,0 +1,2 @@
+The `openmpf_python_component_build` image works in conjunction with the `openmpf_python_executor` image. 
+Documentation for both can be found in [`python_executor/README.md`](../python_executor/README.md).
